test(cli): split RN dev input tests
